requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-cloud-storage==1.27.0
